requirements.txt
pandas==2.1.4
folium==0.15.1
Pillow==10.1.0
openpyxl==3.1.2
Flask==3.0.0
sqlite3 